Increase JobIter channel prefetch
This allows to fetch more jobs to workers while some of the previous
ones has not yet been acknoledged, thus passing the new
TestConsumersCanShareJobIteratorConcurrently.

But this prefetch number should be tunned to the number of workers.

IMHO, a better solution would be to pass a new Channel to every worker
instead, as it is the recommended solution to this problem. See why at
https://www.rabbitmq.com/tutorials/amqp-concepts.html#amqp-channels