Use blocking call with timeout instead of retries
After each retry it waited a fixed amount of time. This caused it to
retry at least once per message, that is, a 50 ms extra sleep per job.

Now it blocks for a configurable time if no messages are arriving. No
retry is needed and jobs are processes as fast as they arrive.

Signed-off-by: Javi Fontan <jfontan@gmail.com>